modified hw4
