+++ src/hw4_Polymorphism/Competition/Pacticipant/Cat.java
