+++ src/hw4_Polymorphism/Figures/Main.java
@@ -20,6 +20,6 @@ public class Main {
         for (Figure figure: figures){
             sum += figure.countArea();
         }
-        return sum;
+        return Math.round(sum);
     }
 }
